jobThrottle now a float; time it takes to submit a job is now in the score feedback loop. by default it is set to stabilize at 20s submission time
git-svn-id: https://svn.code.sf.net/p/cogkit/svn/trunk/current/src/cog@1899 5b74d2a0-fa0e-0410-85ed-ffba77ec0bde

+++ modules/karajan/src/org/globus/cog/karajan/scheduler/WeightedHost.java
@@ -20,17 +20,17 @@ public class WeightedHost implements Comparable {
 	private double tscore;
 	private int load;
 	private double delayedDelta;
-	private int jobThrottle;
+	private float jobThrottle;
 
-	public WeightedHost(BoundContact contact, int jobThrottle) {
+	public WeightedHost(BoundContact contact, float jobThrottle) {
 		this(contact, 0.0, jobThrottle);
 	}
 
-	public WeightedHost(BoundContact contact, double score, int jobThrottle) {
+	public WeightedHost(BoundContact contact, double score, float jobThrottle) {
 		this(contact, score, 0, jobThrottle);
 	}
 
-	public WeightedHost(BoundContact contact, double score, int load, int jobThrottle) {
+	public WeightedHost(BoundContact contact, double score, int load, float jobThrottle) {
 		this.host = contact;
 		setScore(score);
 		this.load = load;
@@ -129,7 +129,7 @@ public class WeightedHost implements Comparable {
 		return !(load < jobThrottle * tscore + 2);
 	}
 
-	public int getJobThrottle() {
+	public float getJobThrottle() {
 		return jobThrottle;
 	}
 }
